# openapi_pathology_hub_unified_searchEvidence_v1_5_9_curriculum.yaml
openapi: 3.1.0
info:
  title: Pathology Hub Unified Evidence API
  version: 1.5.9-curriculum-map-v02
  description: Single-action evidence search for Pathology Hub, extended with Curriculum Map v0.2.
servers:
  - url: https://pathology-hub-v04-curriculum-staging-vorn5q2kga-uc.a.run.app
paths:
  /health:
    get:
      summary: Service health.
      responses:
        "200":
          description: Health response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /evidence/search:
    post:
      operationId: searchEvidence
      summary: Search Pathology Hub evidence sources.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - query
              properties:
                query:
                  type: string
                sources:
                  type: array
                  items:
                    type: string
                    enum:
                      - who
                      - textbooks
                      - journals
                      - pathout
                      - lectures
                      - videos
                      - curriculum
                  default:
                    - textbooks
                max_results:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 3
                include_figures:
                  type: boolean
                  default: false
                max_figures:
                  type: integer
                  minimum: 0
                  maximum: 10
                  default: 0
                compact:
                  type: boolean
                  default: true
                excerpt_char_limit:
                  type: integer
                  minimum: 200
                  maximum: 4000
                  default: 900
      responses:
        "200":
          description: Evidence search response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Bad request.
        "401":
          description: Unauthorized.
        "500":
          description: Server error.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key